requirements.txt
requests==2.28.2
telebot==0.0.5
python-dotenv==1.0.0
loguru==0.6.0
backoff==2.2.1
python-telegram-bot-calendar==1.0.5
PyMySQL==1.0.3
